refactor(ai-plan-import): extrait le modèle domaine du job hors du fichier table drizzle
