+++ apps/backend/src/plans/plans/ai-plan-import/get-import-status/get-import-status.service.ts
@@ -8,7 +8,7 @@ import {
   type AiPlanImportError,
 } from '../ai-plan-import.errors';
 import { AiPlanImportJobRepository } from '../ai-plan-import-job.repository';
-import { AiPlanImportJob } from '../models/ai-plan-import-job.table';
+import { AiPlanImportJob } from '../models/ai-plan-import-job';
 import type { GetImportStatusOutput } from './get-import-status.output';
 
 export type GetImportStatusServiceInput = {
